requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
flasgger==0.9.7.1
SQLAlchemy==2.0.43
Flask-Login==0.6.3
Flask-WTF==1.2.2
email-validator==2.3.0
Werkzeug==3.1.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.3
google-generativeai==0.8.5
google-genai

pandas==2.3.2
openpyxl==3.1.5
celery==5.3.6
redis==5.0.7
APScheduler==3.11.0
bcrypt==4.3.0
python-docx==1.2.0
python-telegram-bot==20.7
python-dotenv==1.0.1
cloudinary==1.36.0
flask-sock
simple-websocket